DAOS-5824 control: move control api print functions into dmg
Control API package has print functions to display human readable
representations of results maps returned from requests made through
the API.  Move these into the pretty package which exists in the
dmg tool directory as the display translation of the API results
should be the responsibility of the API consumer.
