+++ src/control/cmd/daos_server/storage.go
@@ -29,9 +29,9 @@ import (
 
 	"github.com/pkg/errors"
 
+	"github.com/daos-stack/daos/src/control/cmd/dmg/pretty"
 	"github.com/daos-stack/daos/src/control/common"
 	commands "github.com/daos-stack/daos/src/control/common/storage"
-	"github.com/daos-stack/daos/src/control/lib/control"
 	"github.com/daos-stack/daos/src/control/server"
 	"github.com/daos-stack/daos/src/control/server/storage/bdev"
 	"github.com/daos-stack/daos/src/control/server/storage/scm"
@@ -141,7 +141,7 @@ func (cmd *storageScanCmd) Execute(args []string) error {
 		if err != nil {
 			return err
 		}
-		if err := control.PrintNvmeControllers(nvmeResp.Controllers, &bld); err != nil {
+		if err := pretty.PrintNvmeControllers(nvmeResp.Controllers, &bld); err != nil {
 			return err
 		}
 	}
@@ -155,7 +155,7 @@ func (cmd *storageScanCmd) Execute(args []string) error {
 		if err != nil {
 			return err
 		}
-		if err := control.PrintScmNamespaces(scmResp.Namespaces, &bld); err != nil {
+		if err := pretty.PrintScmNamespaces(scmResp.Namespaces, &bld); err != nil {
 			return err
 		}
 	default:
@@ -163,7 +163,7 @@ func (cmd *storageScanCmd) Execute(args []string) error {
 		if err != nil {
 			return err
 		}
-		if err := control.PrintScmModules(scmResp.Modules, &bld); err != nil {
+		if err := pretty.PrintScmModules(scmResp.Modules, &bld); err != nil {
 			return err
 		}
 	}
