+++ src/control/cmd/dmg/network.go
@@ -68,7 +68,7 @@ func (cmd *networkScanCmd) Execute(_ []string) error {
 	}
 
 	var bld strings.Builder
-	if err := control.PrintResponseErrors(resp, &bld); err != nil {
+	if err := pretty.PrintResponseErrors(resp, &bld); err != nil {
 		return err
 	}
