+++ src/control/cmd/dmg/pretty/firmware.go
@@ -125,7 +125,7 @@ type hostDeviceError struct {
 
 // PrintSCMFirmwareQueryMap formats the firmware query results in a condensed format.
 func PrintSCMFirmwareQueryMap(fwMap control.HostSCMQueryMap, out io.Writer,
-	opts ...control.PrintConfigOption) error {
+	opts ...PrintConfigOption) error {
 	if fwMap == nil {
 		return nil
 	}
@@ -143,7 +143,7 @@ func PrintSCMFirmwareQueryMap(fwMap control.HostSCMQueryMap, out io.Writer,
 	}
 
 	return printCondensedResults(successes, iw, opts,
-		func(result string, set *hostDeviceSet, _ []control.PrintConfigOption, w io.Writer) {
+		func(result string, set *hostDeviceSet, _ []PrintConfigOption, w io.Writer) {
 			fmt.Fprintf(w, "Firmware status for %s:\n", english.Plural(len(set.Devices), "device", "devices"))
 
 			iw := txtfmt.NewIndentWriter(w)
@@ -214,15 +214,15 @@ func getPrintVersion(version string) string {
 	return version
 }
 
-func printCondensedResults(condensed hostDeviceResultMap, out io.Writer, opts []control.PrintConfigOption,
-	printResult func(string, *hostDeviceSet, []control.PrintConfigOption, io.Writer)) error {
+func printCondensedResults(condensed hostDeviceResultMap, out io.Writer, opts []PrintConfigOption,
+	printResult func(string, *hostDeviceSet, []PrintConfigOption, io.Writer)) error {
 	w := txtfmt.NewErrWriter(out)
 	for _, result := range condensed.Keys() {
 		set, ok := condensed[result]
 		if !ok {
 			continue
 		}
-		hosts := control.GetPrintHosts(set.Hosts.RangedString(), opts...)
+		hosts := getPrintHosts(set.Hosts.RangedString(), opts...)
 		printHostHeader(hosts, out)
 
 		iw := txtfmt.NewIndentWriter(out)
@@ -243,7 +243,7 @@ func printHostHeader(hosts string, out io.Writer) {
 	fmt.Fprintf(out, "%s\n%s\n%s\n", lineBreak, hosts, lineBreak)
 }
 
-func printDeviceErrorTable(errorResults []hostDeviceError, devTitle string, out io.Writer, opts ...control.PrintConfigOption) error {
+func printDeviceErrorTable(errorResults []hostDeviceError, devTitle string, out io.Writer, opts ...PrintConfigOption) error {
 	if len(errorResults) == 0 {
 		return nil
 	}
@@ -279,7 +279,7 @@ func getShortSCMString(module storage.ScmModule) string {
 
 // PrintSCMFirmwareQueryMapVerbose formats the firmware query results in a detailed format.
 func PrintSCMFirmwareQueryMapVerbose(fwMap control.HostSCMQueryMap, out io.Writer,
-	opts ...control.PrintConfigOption) error {
+	opts ...PrintConfigOption) error {
 	if fwMap == nil {
 		return nil
 	}
@@ -327,7 +327,7 @@ func printSCMFirmwareQueryMapByHost(fwMap control.HostSCMQueryMap, out io.Writer
 
 // PrintSCMFirmwareUpdateMap prints the update results in a condensed format.
 func PrintSCMFirmwareUpdateMap(fwMap control.HostSCMUpdateMap, out io.Writer,
-	opts ...control.PrintConfigOption) error {
+	opts ...PrintConfigOption) error {
 	successes, errs, err := condenseSCMUpdateMap(fwMap)
 	if err != nil {
 		return err
@@ -338,7 +338,7 @@ func PrintSCMFirmwareUpdateMap(fwMap control.HostSCMUpdateMap, out io.Writer,
 	}
 
 	return printCondensedResults(successes, out, opts,
-		func(result string, set *hostDeviceSet, _ []control.PrintConfigOption, w io.Writer) {
+		func(result string, set *hostDeviceSet, _ []PrintConfigOption, w io.Writer) {
 			fmt.Fprintf(w, "Firmware staged on %s. A power cycle is required to apply the update.\n",
 				english.Plural(len(set.Devices), "device", "devices"))
 		})
@@ -379,7 +379,7 @@ func condenseSCMUpdateMap(fwMap control.HostSCMUpdateMap) (hostDeviceResultMap, 
 // PrintSCMFirmwareUpdateMapVerbose formats the firmware update results in a
 // detailed format.
 func PrintSCMFirmwareUpdateMapVerbose(fwMap control.HostSCMUpdateMap, out io.Writer,
-	opts ...control.PrintConfigOption) error {
+	opts ...PrintConfigOption) error {
 	w := txtfmt.NewErrWriter(out)
 
 	for _, host := range fwMap.Keys() {
@@ -414,7 +414,7 @@ func PrintSCMFirmwareUpdateMapVerbose(fwMap control.HostSCMUpdateMap, out io.Wri
 // PrintNVMeFirmwareQueryMap formats the NVMe device firmware query results in a
 // concise format.
 func PrintNVMeFirmwareQueryMap(fwMap control.HostNVMeQueryMap, out io.Writer,
-	opts ...control.PrintConfigOption) error {
+	opts ...PrintConfigOption) error {
 	if fwMap == nil {
 		return nil
 	}
@@ -432,7 +432,7 @@ func PrintNVMeFirmwareQueryMap(fwMap control.HostNVMeQueryMap, out io.Writer,
 	}
 
 	return printCondensedResults(successes, iw, opts,
-		func(result string, set *hostDeviceSet, _ []control.PrintConfigOption, w io.Writer) {
+		func(result string, set *hostDeviceSet, _ []PrintConfigOption, w io.Writer) {
 			fmt.Fprintf(w, "Firmware status for %s:\n", english.Plural(len(set.Devices), "device", "devices"))
 
 			iw := txtfmt.NewIndentWriter(w)
@@ -468,7 +468,7 @@ func getNVMeFirmwareQueryStr(result *control.NVMeQueryResult) string {
 // PrintNVMeFirmwareQueryMapVerbose formats the NVMe device firmware query
 // results in a verbose format.
 func PrintNVMeFirmwareQueryMapVerbose(fwMap control.HostNVMeQueryMap, out io.Writer,
-	opts ...control.PrintConfigOption) error {
+	opts ...PrintConfigOption) error {
 	if fwMap == nil {
 		return nil
 	}
@@ -505,7 +505,7 @@ func printNVMeFirmwareQueryMapByHost(fwMap control.HostNVMeQueryMap, w io.Writer
 // PrintNVMeFirmwareUpdateMap formats the NVMe device firmware update results in
 // a concise format.
 func PrintNVMeFirmwareUpdateMap(fwMap control.HostNVMeUpdateMap, out io.Writer,
-	opts ...control.PrintConfigOption) error {
+	opts ...PrintConfigOption) error {
 	successes, errs, err := condenseNVMeUpdateMap(fwMap)
 	if err != nil {
 		return err
@@ -516,7 +516,7 @@ func PrintNVMeFirmwareUpdateMap(fwMap control.HostNVMeUpdateMap, out io.Writer,
 	}
 
 	return printCondensedResults(successes, out, opts,
-		func(result string, set *hostDeviceSet, _ []control.PrintConfigOption, w io.Writer) {
+		func(result string, set *hostDeviceSet, _ []PrintConfigOption, w io.Writer) {
 			fmt.Fprintf(w, "Firmware updated on %s.\n",
 				english.Plural(len(set.Devices), "NVMe device controller", "NVMe device controllers"))
 		})
@@ -554,8 +554,7 @@ func condenseNVMeUpdateMap(fwMap control.HostNVMeUpdateMap) (hostDeviceResultMap
 }
 
 // PrintNVMeFirmwareUpdateMapVerbose prints a verbose listing of firmware update results.
-func PrintNVMeFirmwareUpdateMapVerbose(fwMap control.HostNVMeUpdateMap, out io.Writer,
-	opts ...control.PrintConfigOption) error {
+func PrintNVMeFirmwareUpdateMapVerbose(fwMap control.HostNVMeUpdateMap, out io.Writer, opts ...PrintConfigOption) error {
 	w := txtfmt.NewErrWriter(out)
 
 	for _, host := range fwMap.Keys() {
