+++ src/control/cmd/dmg/pretty/pool.go
@@ -36,7 +36,7 @@ import (
 
 // PrintPoolQueryResponse generates a human-readable representation of the supplied
 // PoolQueryResp struct and writes it to the supplied io.Writer.
-func PrintPoolQueryResponse(pqr *control.PoolQueryResp, out io.Writer, opts ...control.PrintConfigOption) error {
+func PrintPoolQueryResponse(pqr *control.PoolQueryResp, out io.Writer, opts ...PrintConfigOption) error {
 	if pqr == nil {
 		return errors.Errorf("nil %T", pqr)
 	}
