+++ src/control/cmd/dmg/pretty/storage_test.go
@@ -24,17 +24,18 @@
 package pretty
 
 import (
+	"context"
+	"errors"
 	"fmt"
-	"strconv"
 	"strings"
 	"testing"
 	"time"
 
-	"github.com/dustin/go-humanize"
 	"github.com/google/go-cmp/cmp"
 
 	"github.com/daos-stack/daos/src/control/common"
 	"github.com/daos-stack/daos/src/control/lib/control"
+	"github.com/daos-stack/daos/src/control/logging"
 	"github.com/daos-stack/daos/src/control/server/storage"
 )
 
@@ -55,334 +56,849 @@ func mockHostStorageMap(t *testing.T, hosts ...*mockHostStorage) control.HostSto
 	return hsm
 }
 
-func TestPretty_PrintNVMeHealthMap(t *testing.T) {
+func TestControl_PrintStorageScanResponse(t *testing.T) {
 	var (
-		controllerA    = storage.MockNvmeController(1)
-		controllerB    = storage.MockNvmeController(2)
-		controllerAwTS = storage.MockNvmeController(1)
+		standardScan      = control.MockServerScanResp(t, "standard")
+		withNamespaceScan = control.MockServerScanResp(t, "withNamespace")
+		noNVMEScan        = control.MockServerScanResp(t, "noNVME")
+		noSCMScan         = control.MockServerScanResp(t, "noSCM")
+		noStorageScan     = control.MockServerScanResp(t, "noStorage")
+		scmScanFailed     = control.MockServerScanResp(t, "scmFailed")
+		nvmeScanFailed    = control.MockServerScanResp(t, "nvmeFailed")
+		bothScansFailed   = control.MockServerScanResp(t, "bothFailed")
 	)
-	tt, err := strconv.ParseUint("1405544146", 10, 64)
-	if err != nil {
-		t.Fatal(err)
-	}
-	controllerAwTS.HealthStats.Timestamp = tt
-	ttStr := "2014-07-16 20:55:46 +0000 UTC"
 
 	for name, tc := range map[string]struct {
-		hsm         control.HostStorageMap
+		mic         *control.MockInvokerConfig
 		expPrintStr string
 	}{
-		"no devices": {
-			hsm: mockHostStorageMap(t, &mockHostStorage{"host1", &control.HostStorage{}}),
+		"empty response": {
+			mic: &control.MockInvokerConfig{
+				UnaryResponse: &control.UnaryResponse{},
+			},
+		},
+		"server error": {
+			mic: &control.MockInvokerConfig{
+				UnaryResponse: &control.UnaryResponse{
+					Responses: []*control.HostResponse{
+						{
+							Addr:  "host1",
+							Error: errors.New("failed"),
+						},
+					},
+				},
+			},
 			expPrintStr: `
------
-host1
------
-  No NVMe devices detected
+Errors:
+  Hosts Error  
+  ----- -----  
+  host1 failed 
+
 `,
 		},
-		"1 host; 2 devices": {
-			hsm: mockHostStorageMap(t,
-				&mockHostStorage{
-					"host1",
-					&control.HostStorage{
-						NvmeDevices: storage.NvmeControllers{
-							controllerA,
-							controllerB,
+		"scm scan error": {
+			mic: &control.MockInvokerConfig{
+				UnaryResponse: &control.UnaryResponse{
+					Responses: []*control.HostResponse{
+						{
+							Addr:    "host1",
+							Message: scmScanFailed,
 						},
 					},
 				},
-			),
-			expPrintStr: fmt.Sprintf(`
------
-host1
------
-PCI:%s Model:%s FW:%s Socket:%d Capacity:%s
-  Health Stats:
-    Temperature:%dK(%.02fC)
-    Temperature Warning Duration:%dm0s
-    Temperature Critical Duration:%dm0s
-    Controller Busy Time:%dm0s
-    Power Cycles:%d
-    Power On Duration:%s
-    Unsafe Shutdowns:%d
-    Media Errors:%d
-    Error Log Entries:%d
-  Critical Warnings:
-    Temperature: WARNING
-    Available Spare: WARNING
-    Device Reliability: WARNING
-    Read Only: WARNING
-    Volatile Memory Backup: WARNING
+			},
+			expPrintStr: `
+Errors:
+  Hosts Error           
+  ----- -----           
+  host1 scm scan failed 
 
-PCI:%s Model:%s FW:%s Socket:%d Capacity:%s
-  Health Stats:
-    Temperature:%dK(%.02fC)
-    Temperature Warning Duration:%dm0s
-    Temperature Critical Duration:%dm0s
-    Controller Busy Time:%dm0s
-    Power Cycles:%d
-    Power On Duration:%s
-    Unsafe Shutdowns:%d
-    Media Errors:%d
-    Error Log Entries:%d
-  Critical Warnings:
-    Temperature: WARNING
-    Available Spare: WARNING
-    Device Reliability: WARNING
-    Read Only: WARNING
-    Volatile Memory Backup: WARNING
-
+Hosts SCM Total       NVMe Total         
+----- ---------       ----------         
+host1 0 B (0 modules) 1 B (1 controller) 
 `,
-				controllerA.PciAddr, controllerA.Model, controllerA.FwRev,
-				controllerA.SocketID, humanize.Bytes(controllerA.Capacity()),
-				controllerA.HealthStats.TempK(), controllerA.HealthStats.TempC(),
-				controllerA.HealthStats.TempWarnTime, controllerA.HealthStats.TempCritTime,
-				controllerA.HealthStats.CtrlBusyTime, controllerA.HealthStats.PowerCycles,
-				time.Duration(controllerA.HealthStats.PowerOnHours)*time.Hour,
-				controllerA.HealthStats.UnsafeShutdowns, controllerA.HealthStats.MediaErrors,
-				controllerA.HealthStats.ErrorLogEntries,
-
-				controllerB.PciAddr, controllerB.Model, controllerB.FwRev, controllerB.SocketID,
-				humanize.Bytes(controllerB.Capacity()),
-				controllerB.HealthStats.TempK(), controllerB.HealthStats.TempC(),
-				controllerB.HealthStats.TempWarnTime, controllerB.HealthStats.TempCritTime,
-				controllerB.HealthStats.CtrlBusyTime, controllerB.HealthStats.PowerCycles,
-				time.Duration(controllerB.HealthStats.PowerOnHours)*time.Hour,
-				controllerB.HealthStats.UnsafeShutdowns, controllerB.HealthStats.MediaErrors,
-				controllerB.HealthStats.ErrorLogEntries,
-			),
 		},
-		"1 host; 1 device, fetched over drpc": {
-			hsm: mockHostStorageMap(t,
-				&mockHostStorage{
-					"host1",
-					&control.HostStorage{
-						NvmeDevices: storage.NvmeControllers{
-							controllerAwTS,
+		"nvme scan error": {
+			mic: &control.MockInvokerConfig{
+				UnaryResponse: &control.UnaryResponse{
+					Responses: []*control.HostResponse{
+						{
+							Addr:    "host1",
+							Message: nvmeScanFailed,
 						},
 					},
 				},
-			),
-			expPrintStr: fmt.Sprintf(`
------
-host1
------
-PCI:%s Model:%s FW:%s Socket:%d Capacity:%s
-  Health Stats:
-    Timestamp:%s
-    Temperature:%dK(%.02fC)
-    Temperature Warning Duration:%dm0s
-    Temperature Critical Duration:%dm0s
-    Controller Busy Time:%dm0s
-    Power Cycles:%d
-    Power On Duration:%s
-    Unsafe Shutdowns:%d
-    Media Errors:%d
-    Read Errors:%d
-    Write Errors:%d
-    Unmap Errors:%d
-    Checksum Errors:%d
-    Error Log Entries:%d
-  Critical Warnings:
-    Temperature: WARNING
-    Available Spare: WARNING
-    Device Reliability: WARNING
-    Read Only: WARNING
-    Volatile Memory Backup: WARNING
+			},
+			expPrintStr: `
+Errors:
+  Hosts Error            
+  ----- -----            
+  host1 nvme scan failed 
 
+Hosts SCM Total      NVMe Total          
+----- ---------      ----------          
+host1 1 B (1 module) 0 B (0 controllers) 
 `,
-				controllerAwTS.PciAddr, controllerAwTS.Model, controllerAwTS.FwRev,
-				controllerAwTS.SocketID, humanize.Bytes(controllerAwTS.Capacity()), ttStr,
-				controllerAwTS.HealthStats.TempK(), controllerAwTS.HealthStats.TempC(),
-				controllerAwTS.HealthStats.TempWarnTime, controllerAwTS.HealthStats.TempCritTime,
-				controllerAwTS.HealthStats.CtrlBusyTime, controllerAwTS.HealthStats.PowerCycles,
-				time.Duration(controllerAwTS.HealthStats.PowerOnHours)*time.Hour,
-				controllerAwTS.HealthStats.UnsafeShutdowns, controllerAwTS.HealthStats.MediaErrors,
-				controllerAwTS.HealthStats.ReadErrors, controllerAwTS.HealthStats.WriteErrors,
-				controllerAwTS.HealthStats.UnmapErrors, controllerAwTS.HealthStats.ChecksumErrors,
-				controllerAwTS.HealthStats.ErrorLogEntries,
-			),
+		},
+		"scm and nvme scan error": {
+			mic: &control.MockInvokerConfig{
+				UnaryResponse: &control.UnaryResponse{
+					Responses: []*control.HostResponse{
+						{
+							Addr:    "host1:1",
+							Message: bothScansFailed,
+						},
+						{
+							Addr:    "host2:1",
+							Message: bothScansFailed,
+						},
+					},
+				},
+			},
+			expPrintStr: `
+Errors:
+  Hosts     Error            
+  -----     -----            
+  host[1-2] nvme scan failed 
+  host[1-2] scm scan failed  
+
+Hosts     SCM Total       NVMe Total          
+-----     ---------       ----------          
+host[1-2] 0 B (0 modules) 0 B (0 controllers) 
+`,
+		},
+		"no storage": {
+			mic: &control.MockInvokerConfig{
+				UnaryResponse: &control.UnaryResponse{
+					Responses: []*control.HostResponse{
+						{
+							Addr:    "host1",
+							Message: noStorageScan,
+						},
+					},
+				},
+			},
+			expPrintStr: `
+Hosts SCM Total       NVMe Total          
+----- ---------       ----------          
+host1 0 B (0 modules) 0 B (0 controllers) 
+`,
+		},
+		"single host": {
+			mic: &control.MockInvokerConfig{
+				UnaryResponse: &control.UnaryResponse{
+					Responses: []*control.HostResponse{
+						{
+							Addr:    "host1",
+							Message: standardScan,
+						},
+					},
+				},
+			},
+			expPrintStr: `
+Hosts SCM Total      NVMe Total         
+----- ---------      ----------         
+host1 1 B (1 module) 1 B (1 controller) 
+`,
+		},
+		"single host with namespace": {
+			mic: &control.MockInvokerConfig{
+				UnaryResponse: &control.UnaryResponse{
+					Responses: []*control.HostResponse{
+						{
+							Addr:    "host1",
+							Message: withNamespaceScan,
+						},
+					},
+				},
+			},
+			expPrintStr: `
+Hosts SCM Total         NVMe Total         
+----- ---------         ----------         
+host1 1 B (1 namespace) 1 B (1 controller) 
+`,
+		},
+		"two hosts same scan": {
+			mic: &control.MockInvokerConfig{
+				UnaryResponse: &control.UnaryResponse{
+					Responses: []*control.HostResponse{
+						{
+							Addr:    "host1",
+							Message: standardScan,
+						},
+						{
+							Addr:    "host2",
+							Message: standardScan,
+						},
+					},
+				},
+			},
+			expPrintStr: `
+Hosts     SCM Total      NVMe Total         
+-----     ---------      ----------         
+host[1-2] 1 B (1 module) 1 B (1 controller) 
+`,
+		},
+		"two hosts different scans": {
+			mic: &control.MockInvokerConfig{
+				UnaryResponse: &control.UnaryResponse{
+					Responses: []*control.HostResponse{
+						{
+							Addr:    "host1",
+							Message: noNVMEScan,
+						},
+						{
+							Addr:    "host2",
+							Message: noSCMScan,
+						},
+					},
+				},
+			},
+			expPrintStr: `
+Hosts SCM Total       NVMe Total          
+----- ---------       ----------          
+host1 1 B (1 module)  0 B (0 controllers) 
+host2 0 B (0 modules) 1 B (1 controller)  
+`,
+		},
+		"1024 hosts same scan": {
+			mic: &control.MockInvokerConfig{
+				UnaryResponse: &control.UnaryResponse{
+					Responses: control.MockHostResponses(t,
+						1024, "host%000d", standardScan),
+				},
+			},
+			expPrintStr: `
+Hosts        SCM Total      NVMe Total         
+-----        ---------      ----------         
+host[0-1023] 1 B (1 module) 1 B (1 controller) 
+`,
 		},
 	} {
 		t.Run(name, func(t *testing.T) {
+			log, buf := logging.NewTestLogger(t.Name())
+			defer common.ShowBufferOnFailure(t, buf)
+
+			ctx := context.TODO()
+			mi := control.NewMockInvoker(log, tc.mic)
+
+			resp, err := control.StorageScan(ctx, mi, &control.StorageScanReq{})
+			if err != nil {
+				t.Fatal(err)
+			}
+
 			var bld strings.Builder
-			if err := PrintNvmeHealthMap(tc.hsm, &bld); err != nil {
+			if err := PrintResponseErrors(resp, &bld); err != nil {
+				t.Fatal(err)
+			}
+			if err := PrintHostStorageMap(resp.HostStorage, &bld); err != nil {
 				t.Fatal(err)
 			}
 
 			if diff := cmp.Diff(strings.TrimLeft(tc.expPrintStr, "\n"), bld.String()); diff != "" {
-				t.Fatalf("unexpected print output (-want, +got):\n%s\n", diff)
+				t.Fatalf("unexpected format string (-want, +got):\n%s\n", diff)
 			}
 		})
 	}
 }
 
-func TestPretty_PrintNVMetaMap(t *testing.T) {
+func TestControl_PrintStorageScanResponseVerbose(t *testing.T) {
 	var (
-		controllerA = storage.MockNvmeController(1)
-		controllerB = storage.MockNvmeController(2)
-		controllerC = storage.MockNvmeController(1)
-		controllerD = storage.MockNvmeController(2)
-		controllerE = storage.MockNvmeController(1)
-		controllerF = storage.MockNvmeController(2)
+		standardScan      = control.MockServerScanResp(t, "standard")
+		withNamespaceScan = control.MockServerScanResp(t, "withNamespace")
+		noNVMEScan        = control.MockServerScanResp(t, "noNVME")
+		noSCMScan         = control.MockServerScanResp(t, "noSCM")
+		noStorageScan     = control.MockServerScanResp(t, "noStorage")
+		scmScanFailed     = control.MockServerScanResp(t, "scmFailed")
+		nvmeScanFailed    = control.MockServerScanResp(t, "nvmeFailed")
+		bothScansFailed   = control.MockServerScanResp(t, "bothFailed")
 	)
-	controllerA.SmdDevices = nil
-	controllerB.SmdDevices = nil
-	controllerE.SmdDevices = []*storage.SmdDevice{
-		{
-			UUID:      common.MockUUID(0),
-			TargetIDs: []int32{0, 1, 2},
-			Rank:      0,
-			State:     "NORMAL",
-		},
-		{
-			UUID:      common.MockUUID(1),
-			TargetIDs: []int32{3, 4, 5},
-			Rank:      0,
-			State:     "FAULTY",
-		},
-	}
-	controllerF.SmdDevices = []*storage.SmdDevice{
-		{
-			UUID:      common.MockUUID(2),
-			TargetIDs: []int32{6, 7, 8},
-			Rank:      1,
-			State:     "NORMAL",
-		},
-		{
-			UUID:      common.MockUUID(3),
-			TargetIDs: []int32{9, 10, 11},
-			Rank:      1,
-			State:     "FAULTY",
-		},
-	}
+
 	for name, tc := range map[string]struct {
-		hsm         control.HostStorageMap
+		mic         *control.MockInvokerConfig
 		expPrintStr string
 	}{
-		"no controllers": {
-			hsm: mockHostStorageMap(t, &mockHostStorage{"host1", &control.HostStorage{}}),
+		"empty response": {
+			mic: &control.MockInvokerConfig{
+				UnaryResponse: &control.UnaryResponse{},
+			},
+		},
+		"server error": {
+			mic: &control.MockInvokerConfig{
+				UnaryResponse: &control.UnaryResponse{
+					Responses: []*control.HostResponse{
+						{
+							Addr:  "host1",
+							Error: errors.New("failed"),
+						},
+					},
+				},
+			},
+			expPrintStr: `
+Errors:
+  Hosts Error  
+  ----- -----  
+  host1 failed 
+
+`,
+		},
+		"scm scan error": {
+			mic: &control.MockInvokerConfig{
+				UnaryResponse: &control.UnaryResponse{
+					Responses: []*control.HostResponse{
+						{
+							Addr:    "host1",
+							Message: scmScanFailed,
+						},
+					},
+				},
+			},
+			expPrintStr: `
+Errors:
+  Hosts Error           
+  ----- -----           
+  host1 scm scan failed 
+
+-----
+host1
+-----
+	No SCM modules found
+
+NVMe PCI     Model   FW Revision Socket ID Capacity 
+--------     -----   ----------- --------- -------- 
+0000:80:00.1 model-1 fwRev-1     1         1 B      
+
+`,
+		},
+		"nvme scan error": {
+			mic: &control.MockInvokerConfig{
+				UnaryResponse: &control.UnaryResponse{
+					Responses: []*control.HostResponse{
+						{
+							Addr:    "host1",
+							Message: nvmeScanFailed,
+						},
+					},
+				},
+			},
+			expPrintStr: `
+Errors:
+  Hosts Error            
+  ----- -----            
+  host1 nvme scan failed 
+
+-----
+host1
+-----
+SCM Module ID Socket ID Memory Ctrlr ID Channel ID Channel Slot Capacity 
+------------- --------- --------------- ---------- ------------ -------- 
+1             1         1               1          1            1 B      
+
+	No NVMe devices found
+
+`,
+		},
+		"scm and nvme scan error": {
+			mic: &control.MockInvokerConfig{
+				UnaryResponse: &control.UnaryResponse{
+					Responses: []*control.HostResponse{
+						{
+							Addr:    "host1:1",
+							Message: bothScansFailed,
+						},
+						{
+							Addr:    "host2:1",
+							Message: bothScansFailed,
+						},
+					},
+				},
+			},
+			expPrintStr: `
+Errors:
+  Hosts     Error            
+  -----     -----            
+  host[1-2] nvme scan failed 
+  host[1-2] scm scan failed  
+
+---------
+host[1-2]
+---------
+	No SCM modules found
+
+	No NVMe devices found
+
+`,
+		},
+		"no storage": {
+			mic: &control.MockInvokerConfig{
+				UnaryResponse: &control.UnaryResponse{
+					Responses: []*control.HostResponse{
+						{
+							Addr:    "host1",
+							Message: noStorageScan,
+						},
+						{
+							Addr:    "host2",
+							Message: noStorageScan,
+						},
+					},
+				},
+			},
+			expPrintStr: `
+---------
+host[1-2]
+---------
+	No SCM modules found
+
+	No NVMe devices found
+
+`,
+		},
+		"single host": {
+			mic: &control.MockInvokerConfig{
+				UnaryResponse: &control.UnaryResponse{
+					Responses: []*control.HostResponse{
+						{
+							Addr:    "host1",
+							Message: standardScan,
+						},
+					},
+				},
+			},
 			expPrintStr: `
 -----
 host1
 -----
-  No NVMe devices detected
+SCM Module ID Socket ID Memory Ctrlr ID Channel ID Channel Slot Capacity 
+------------- --------- --------------- ---------- ------------ -------- 
+1             1         1               1          1            1 B      
+
+NVMe PCI     Model   FW Revision Socket ID Capacity 
+--------     -----   ----------- --------- -------- 
+0000:80:00.1 model-1 fwRev-1     1         1 B      
+
 `,
 		},
-		"no smd devices on controllers": {
-			hsm: mockHostStorageMap(t,
-				&mockHostStorage{
-					"host1",
-					&control.HostStorage{
-						NvmeDevices: storage.NvmeControllers{
-							controllerA,
-							controllerB,
+		"single host with namespace": {
+			mic: &control.MockInvokerConfig{
+				UnaryResponse: &control.UnaryResponse{
+					Responses: []*control.HostResponse{
+						{
+							Addr:    "host1",
+							Message: withNamespaceScan,
 						},
 					},
 				},
-			),
-			expPrintStr: fmt.Sprintf(`
+			},
+			expPrintStr: `
 -----
 host1
 -----
-PCI:%s Model:%s FW:%s Socket:%d Capacity:%s
-  No SMD devices found
+SCM Namespace Socket ID Capacity 
+------------- --------- -------- 
+/dev/pmem1    1         1 B      
 
-PCI:%s Model:%s FW:%s Socket:%d Capacity:%s
-  No SMD devices found
+NVMe PCI     Model   FW Revision Socket ID Capacity 
+--------     -----   ----------- --------- -------- 
+0000:80:00.1 model-1 fwRev-1     1         1 B      
 
 `,
-				controllerA.PciAddr, controllerA.Model, controllerA.FwRev,
-				controllerA.SocketID, humanize.Bytes(controllerA.Capacity()),
-				controllerB.PciAddr, controllerB.Model, controllerB.FwRev,
-				controllerB.SocketID, humanize.Bytes(controllerB.Capacity())),
 		},
-		"single smd device on each controller": {
-			hsm: mockHostStorageMap(t,
-				&mockHostStorage{
-					"host1",
-					&control.HostStorage{
-						NvmeDevices: storage.NvmeControllers{
-							controllerC,
-							controllerD,
+		"two hosts same scan": {
+			mic: &control.MockInvokerConfig{
+				UnaryResponse: &control.UnaryResponse{
+					Responses: []*control.HostResponse{
+						{
+							Addr:    "host1",
+							Message: standardScan,
+						},
+						{
+							Addr:    "host2",
+							Message: standardScan,
 						},
 					},
 				},
-			),
-			expPrintStr: fmt.Sprintf(`
------
-host1
------
-PCI:%s Model:%s FW:%s Socket:%d Capacity:%s
-  SMD Devices
-    UUID:%s Targets:%v Rank:%d State:%s
+			},
+			expPrintStr: `
+---------
+host[1-2]
+---------
+SCM Module ID Socket ID Memory Ctrlr ID Channel ID Channel Slot Capacity 
+------------- --------- --------------- ---------- ------------ -------- 
+1             1         1               1          1            1 B      
 
-PCI:%s Model:%s FW:%s Socket:%d Capacity:%s
-  SMD Devices
-    UUID:%s Targets:%v Rank:%d State:%s
+NVMe PCI     Model   FW Revision Socket ID Capacity 
+--------     -----   ----------- --------- -------- 
+0000:80:00.1 model-1 fwRev-1     1         1 B      
 
 `,
-				controllerC.PciAddr, controllerC.Model, controllerC.FwRev,
-				controllerC.SocketID, humanize.Bytes(controllerC.Capacity()),
-				controllerC.SmdDevices[0].UUID, controllerC.SmdDevices[0].TargetIDs,
-				controllerC.SmdDevices[0].Rank, controllerC.SmdDevices[0].State,
-
-				controllerD.PciAddr, controllerD.Model, controllerD.FwRev,
-				controllerD.SocketID, humanize.Bytes(controllerD.Capacity()),
-				controllerD.SmdDevices[0].UUID, controllerD.SmdDevices[0].TargetIDs,
-				controllerD.SmdDevices[0].Rank, controllerD.SmdDevices[0].State),
 		},
-		"multiple smd devices on each controller": {
-			hsm: mockHostStorageMap(t,
-				&mockHostStorage{
-					"host1",
-					&control.HostStorage{
-						NvmeDevices: storage.NvmeControllers{
-							controllerE,
-							controllerF,
+		"two hosts different scans": {
+			mic: &control.MockInvokerConfig{
+				UnaryResponse: &control.UnaryResponse{
+					Responses: []*control.HostResponse{
+						{
+							Addr:    "host1",
+							Message: noNVMEScan,
+						},
+						{
+							Addr:    "host2",
+							Message: noSCMScan,
 						},
 					},
 				},
-			),
-			expPrintStr: fmt.Sprintf(`
+			},
+			expPrintStr: `
 -----
 host1
 -----
-PCI:%s Model:%s FW:%s Socket:%d Capacity:%s
-  SMD Devices
-    UUID:%s Targets:%v Rank:%d State:%s
-    UUID:%s Targets:%v Rank:%d State:%s
+SCM Module ID Socket ID Memory Ctrlr ID Channel ID Channel Slot Capacity 
+------------- --------- --------------- ---------- ------------ -------- 
+1             1         1               1          1            1 B      
 
-PCI:%s Model:%s FW:%s Socket:%d Capacity:%s
-  SMD Devices
-    UUID:%s Targets:%v Rank:%d State:%s
-    UUID:%s Targets:%v Rank:%d State:%s
+	No NVMe devices found
+
+-----
+host2
+-----
+	No SCM modules found
+
+NVMe PCI     Model   FW Revision Socket ID Capacity 
+--------     -----   ----------- --------- -------- 
+0000:80:00.1 model-1 fwRev-1     1         1 B      
 
 `,
-				controllerE.PciAddr, controllerE.Model, controllerE.FwRev,
-				controllerE.SocketID, humanize.Bytes(controllerE.Capacity()),
-				controllerE.SmdDevices[0].UUID, controllerE.SmdDevices[0].TargetIDs,
-				controllerE.SmdDevices[0].Rank, controllerE.SmdDevices[0].State,
-				controllerE.SmdDevices[1].UUID, controllerE.SmdDevices[1].TargetIDs,
-				controllerE.SmdDevices[1].Rank, controllerE.SmdDevices[1].State,
+		},
+		"1024 hosts same scan": {
+			mic: &control.MockInvokerConfig{
+				UnaryResponse: &control.UnaryResponse{
+					Responses: control.MockHostResponses(t,
+						1024, "host%000d", standardScan),
+				},
+			},
+			expPrintStr: `
+------------
+host[0-1023]
+------------
+SCM Module ID Socket ID Memory Ctrlr ID Channel ID Channel Slot Capacity 
+------------- --------- --------------- ---------- ------------ -------- 
+1             1         1               1          1            1 B      
 
-				controllerF.PciAddr, controllerF.Model, controllerF.FwRev,
-				controllerF.SocketID, humanize.Bytes(controllerF.Capacity()),
-				controllerF.SmdDevices[0].UUID, controllerF.SmdDevices[0].TargetIDs,
-				controllerF.SmdDevices[0].Rank, controllerF.SmdDevices[0].State,
-				controllerF.SmdDevices[1].UUID, controllerF.SmdDevices[1].TargetIDs,
-				controllerF.SmdDevices[1].Rank, controllerF.SmdDevices[1].State),
+NVMe PCI     Model   FW Revision Socket ID Capacity 
+--------     -----   ----------- --------- -------- 
+0000:80:00.1 model-1 fwRev-1     1         1 B      
+
+`,
 		},
 	} {
 		t.Run(name, func(t *testing.T) {
+			log, buf := logging.NewTestLogger(t.Name())
+			defer common.ShowBufferOnFailure(t, buf)
+
+			ctx := context.TODO()
+			mi := control.NewMockInvoker(log, tc.mic)
+
+			resp, err := control.StorageScan(ctx, mi, &control.StorageScanReq{})
+			if err != nil {
+				t.Fatal(err)
+			}
+
 			var bld strings.Builder
-			if err := PrintNvmeMetaMap(tc.hsm, &bld); err != nil {
+			if err := PrintResponseErrors(resp, &bld); err != nil {
+				t.Fatal(err)
+			}
+			if err := PrintHostStorageMap(resp.HostStorage, &bld, PrintWithVerboseOutput(true)); err != nil {
 				t.Fatal(err)
 			}
 
 			if diff := cmp.Diff(strings.TrimLeft(tc.expPrintStr, "\n"), bld.String()); diff != "" {
-				t.Fatalf("unexpected print output (-want, +got):\n%s\n", diff)
+				t.Fatalf("unexpected format string (-want, +got):\n%s\n", diff)
+			}
+		})
+	}
+}
+
+func TestControl_PrintStorageFormatResponse(t *testing.T) {
+	for name, tc := range map[string]struct {
+		resp        *control.StorageFormatResp
+		expPrintStr string
+	}{
+		"empty response": {
+			resp: &control.StorageFormatResp{},
+		},
+		"server error": {
+			resp: &control.StorageFormatResp{
+				HostErrorsResp: control.MockHostErrorsResp(t,
+					&control.MockHostError{"host1", "failed"}),
+			},
+			expPrintStr: `
+Errors:
+  Hosts Error  
+  ----- -----  
+  host1 failed 
+
+`,
+		},
+		"2 SCM, 2 NVMe; first SCM fails": {
+			resp: control.MockFormatResp(t, control.MockFormatConf{
+				Hosts:       1,
+				ScmPerHost:  2,
+				ScmFailures: control.MockFailureMap(0),
+				NvmePerHost: 2,
+			}),
+			expPrintStr: `
+Errors:
+  Hosts Error                
+  ----- -----                
+  host1 /mnt/1 format failed 
+
+Format Summary:
+  Hosts SCM Devices NVMe Devices 
+  ----- ----------- ------------ 
+  host1 1           1            
+`,
+		},
+		"2 SCM, 2 NVMe; second NVMe fails": {
+			resp: control.MockFormatResp(t, control.MockFormatConf{
+				Hosts:        1,
+				ScmPerHost:   2,
+				NvmePerHost:  2,
+				NvmeFailures: control.MockFailureMap(1),
+			}),
+			expPrintStr: `
+Errors:
+  Hosts Error                       
+  ----- -----                       
+  host1 NVMe device 2 format failed 
+
+Format Summary:
+  Hosts SCM Devices NVMe Devices 
+  ----- ----------- ------------ 
+  host1 2           1            
+`,
+		},
+		"2 SCM, 2 NVMe": {
+			resp: control.MockFormatResp(t, control.MockFormatConf{
+				Hosts:       1,
+				ScmPerHost:  2,
+				NvmePerHost: 2,
+			}),
+			expPrintStr: `
+
+Format Summary:
+  Hosts SCM Devices NVMe Devices 
+  ----- ----------- ------------ 
+  host1 2           2            
+`,
+		},
+		"2 Hosts, 2 SCM, 2 NVMe; first SCM fails": {
+			resp: control.MockFormatResp(t, control.MockFormatConf{
+				Hosts:       2,
+				ScmPerHost:  2,
+				ScmFailures: control.MockFailureMap(0),
+				NvmePerHost: 2,
+			}),
+			expPrintStr: `
+Errors:
+  Hosts     Error                
+  -----     -----                
+  host[1-2] /mnt/1 format failed 
+
+Format Summary:
+  Hosts     SCM Devices NVMe Devices 
+  -----     ----------- ------------ 
+  host[1-2] 1           1            
+`,
+		},
+		"2 Hosts, 2 SCM, 2 NVMe": {
+			resp: control.MockFormatResp(t, control.MockFormatConf{
+				Hosts:       2,
+				ScmPerHost:  2,
+				NvmePerHost: 2,
+			}),
+			expPrintStr: `
+
+Format Summary:
+  Hosts     SCM Devices NVMe Devices 
+  -----     ----------- ------------ 
+  host[1-2] 2           2            
+`,
+		},
+	} {
+		t.Run(name, func(t *testing.T) {
+			var bld strings.Builder
+			if err := PrintResponseErrors(tc.resp, &bld); err != nil {
+				t.Fatal(err)
+			}
+			if err := PrintStorageFormatMap(tc.resp.HostStorage, &bld); err != nil {
+				t.Fatal(err)
+			}
+
+			if diff := cmp.Diff(strings.TrimLeft(tc.expPrintStr, "\n"), bld.String()); diff != "" {
+				t.Fatalf("unexpected format string (-want, +got):\n%s\n", diff)
+			}
+		})
+	}
+}
+
+func TestControl_PrintStorageFormatResponseVerbose(t *testing.T) {
+	for name, tc := range map[string]struct {
+		resp        *control.StorageFormatResp
+		expPrintStr string
+	}{
+		"empty response": {
+			resp: &control.StorageFormatResp{},
+		},
+		"server error": {
+			resp: &control.StorageFormatResp{
+				HostErrorsResp: control.MockHostErrorsResp(t,
+					&control.MockHostError{"host1", "failed"}),
+			},
+			expPrintStr: `
+Errors:
+  Hosts Error  
+  ----- -----  
+  host1 failed 
+
+`,
+		},
+		"2 SCM, 2 NVMe; first SCM fails": {
+			resp: control.MockFormatResp(t, control.MockFormatConf{
+				Hosts:       1,
+				ScmPerHost:  2,
+				ScmFailures: control.MockFailureMap(0),
+				NvmePerHost: 2,
+			}),
+			expPrintStr: `
+Errors:
+  Hosts Error                
+  ----- -----                
+  host1 /mnt/1 format failed 
+
+-----
+host1
+-----
+SCM Mount Format Result 
+--------- ------------- 
+/mnt/2    CTL_SUCCESS   
+
+NVMe PCI Format Result 
+-------- ------------- 
+2        CTL_SUCCESS   
+
+`,
+		},
+		"2 SCM, 2 NVMe; second NVMe fails": {
+			resp: control.MockFormatResp(t, control.MockFormatConf{
+				Hosts:        1,
+				ScmPerHost:   2,
+				NvmePerHost:  2,
+				NvmeFailures: control.MockFailureMap(1),
+			}),
+			expPrintStr: `
+Errors:
+  Hosts Error                       
+  ----- -----                       
+  host1 NVMe device 2 format failed 
+
+-----
+host1
+-----
+SCM Mount Format Result 
+--------- ------------- 
+/mnt/1    CTL_SUCCESS   
+/mnt/2    CTL_SUCCESS   
+
+NVMe PCI Format Result 
+-------- ------------- 
+1        CTL_SUCCESS   
+
+`,
+		},
+		"2 SCM, 2 NVMe": {
+			resp: control.MockFormatResp(t, control.MockFormatConf{
+				Hosts:       1,
+				ScmPerHost:  2,
+				NvmePerHost: 2,
+			}),
+			expPrintStr: `
+
+-----
+host1
+-----
+SCM Mount Format Result 
+--------- ------------- 
+/mnt/1    CTL_SUCCESS   
+/mnt/2    CTL_SUCCESS   
+
+NVMe PCI Format Result 
+-------- ------------- 
+1        CTL_SUCCESS   
+2        CTL_SUCCESS   
+
+`,
+		},
+		"2 Hosts, 2 SCM, 2 NVMe; first SCM fails": {
+			resp: control.MockFormatResp(t, control.MockFormatConf{
+				Hosts:       2,
+				ScmPerHost:  2,
+				ScmFailures: control.MockFailureMap(0),
+				NvmePerHost: 2,
+			}),
+			expPrintStr: `
+Errors:
+  Hosts     Error                
+  -----     -----                
+  host[1-2] /mnt/1 format failed 
+
+---------
+host[1-2]
+---------
+SCM Mount Format Result 
+--------- ------------- 
+/mnt/2    CTL_SUCCESS   
+
+NVMe PCI Format Result 
+-------- ------------- 
+2        CTL_SUCCESS   
+
+`,
+		},
+		"2 Hosts, 2 SCM, 2 NVMe": {
+			resp: control.MockFormatResp(t, control.MockFormatConf{
+				Hosts:       2,
+				ScmPerHost:  2,
+				NvmePerHost: 2,
+			}),
+			expPrintStr: `
+
+---------
+host[1-2]
+---------
+SCM Mount Format Result 
+--------- ------------- 
+/mnt/1    CTL_SUCCESS   
+/mnt/2    CTL_SUCCESS   
+
+NVMe PCI Format Result 
+-------- ------------- 
+1        CTL_SUCCESS   
+2        CTL_SUCCESS   
+
+`,
+		},
+	} {
+		t.Run(name, func(t *testing.T) {
+			var bld strings.Builder
+			if err := PrintResponseErrors(tc.resp, &bld); err != nil {
+				t.Fatal(err)
+			}
+			if err := PrintStorageFormatMap(tc.resp.HostStorage, &bld, PrintWithVerboseOutput(true)); err != nil {
+				t.Fatal(err)
+			}
+
+			if diff := cmp.Diff(strings.TrimLeft(tc.expPrintStr, "\n"), bld.String()); diff != "" {
+				t.Fatalf("unexpected format string (-want, +got):\n%s\n", diff)
 			}
 		})
 	}
@@ -394,7 +910,7 @@ func TestPretty_PrintSmdInfoMap(t *testing.T) {
 	for name, tc := range map[string]struct {
 		req         *control.SmdQueryReq
 		hsm         control.HostStorageMap
-		opts        []control.PrintConfigOption
+		opts        []PrintConfigOption
 		expPrintStr string
 	}{
 		"list-pools (standard)": {
@@ -441,7 +957,7 @@ host1
 			req: &control.SmdQueryReq{
 				OmitDevices: true,
 			},
-			opts: []control.PrintConfigOption{control.PrintWithVerboseOutput(true)},
+			opts: []PrintConfigOption{PrintWithVerboseOutput(true)},
 			hsm: mockHostStorageMap(t,
 				&mockHostStorage{
 					"host1",
