+++ src/control/lib/control/mocks.go
@@ -25,12 +25,16 @@ package control
 
 import (
 	"context"
+	"fmt"
+	"testing"
 
 	"github.com/golang/protobuf/proto"
-)
+	"github.com/pkg/errors"
 
-const (
-	MockUUID = "00000000-1111-2222-3333-444444444444"
+	"github.com/daos-stack/daos/src/control/common/proto/convert"
+	ctlpb "github.com/daos-stack/daos/src/control/common/proto/ctl"
+	"github.com/daos-stack/daos/src/control/lib/hostlist"
+	"github.com/daos-stack/daos/src/control/server/storage"
 )
 
 // MockMessage implements the proto.Message
@@ -128,8 +132,249 @@ func NewMockInvoker(log debugLogger, cfg *MockInvokerConfig) *MockInvoker {
 	}
 }
 
-// DefaultMockInvoker returns a MockInvoker that uses
-// the default configuration.
+// DefaultMockInvoker returns a MockInvoker that uses the default configuration.
 func DefaultMockInvoker(log debugLogger) *MockInvoker {
 	return NewMockInvoker(log, nil)
+}
+
+// MockHostError represents an error received from multiple hosts.
+type MockHostError struct {
+	Hosts string
+	Error string
+}
+
+func mockHostErrorsMap(t *testing.T, hostErrors ...*MockHostError) HostErrorsMap {
+	hem := make(HostErrorsMap)
+
+	for _, he := range hostErrors {
+		if hes, found := hem[he.Error]; found {
+			if _, err := hes.HostSet.Insert(he.Hosts); err != nil {
+				t.Fatal(err)
+			}
+			continue
+		}
+		hem[he.Error] = &HostErrorSet{
+			HostError: errors.New(he.Error),
+			HostSet:   mockHostSet(t, he.Hosts),
+		}
+	}
+
+	return hem
+}
+
+// MockHostErrorsResp returns HostErrorsResp when provided with MockHostErrors
+// from different hostsets.
+func MockHostErrorsResp(t *testing.T, hostErrors ...*MockHostError) HostErrorsResp {
+	if len(hostErrors) == 0 {
+		return HostErrorsResp{}
+	}
+	return HostErrorsResp{
+		HostErrors: mockHostErrorsMap(t, hostErrors...),
+	}
+}
+
+func mockHostSet(t *testing.T, hosts string) *hostlist.HostSet {
+	hs, err := hostlist.CreateSet(hosts)
+	if err != nil {
+		t.Fatal(err)
+	}
+	return hs
+}
+
+func mockHostStorageSet(t *testing.T, hosts string, pbResp *ctlpb.StorageScanResp) *HostStorageSet {
+	hss := &HostStorageSet{
+		HostStorage: &HostStorage{},
+		HostSet:     mockHostSet(t, hosts),
+	}
+
+	if err := convert.Types(pbResp.GetNvme().GetCtrlrs(), &hss.HostStorage.NvmeDevices); err != nil {
+		t.Fatal(err)
+	}
+	if err := convert.Types(pbResp.GetScm().GetModules(), &hss.HostStorage.ScmModules); err != nil {
+		t.Fatal(err)
+	}
+	if err := convert.Types(pbResp.GetScm().GetNamespaces(), &hss.HostStorage.ScmNamespaces); err != nil {
+		t.Fatal(err)
+	}
+
+	return hss
+}
+
+// MockStorageScan represents results from scan on multiple hosts.
+type MockStorageScan struct {
+	Hosts    string
+	HostScan *ctlpb.StorageScanResp
+}
+
+// MockHostStorageMap returns HostStorageMap when provided with mock storage
+// scan results from different hostsets.
+func MockHostStorageMap(t *testing.T, scans ...*MockStorageScan) HostStorageMap {
+	hsm := make(HostStorageMap)
+
+	for _, scan := range scans {
+		hss := mockHostStorageSet(t, scan.Hosts, scan.HostScan)
+		hk, err := hss.HostStorage.HashKey()
+		if err != nil {
+			t.Fatal(err)
+		}
+		hsm[hk] = hss
+	}
+
+	return hsm
+}
+
+func standardServerScanResponse(t *testing.T) *ctlpb.StorageScanResp {
+	pbSsr := &ctlpb.StorageScanResp{
+		Nvme: &ctlpb.ScanNvmeResp{},
+		Scm:  &ctlpb.ScanScmResp{},
+	}
+	nvmeControllers := storage.NvmeControllers{
+		storage.MockNvmeController(),
+	}
+	scmModules := storage.ScmModules{
+		storage.MockScmModule(),
+	}
+	if err := convert.Types(nvmeControllers, &pbSsr.Nvme.Ctrlrs); err != nil {
+		t.Fatal(err)
+	}
+	if err := convert.Types(scmModules, &pbSsr.Scm.Modules); err != nil {
+		t.Fatal(err)
+	}
+
+	return pbSsr
+}
+
+// MocMockServerScanResp returns protobuf storage scan response with contents
+// defined by the variant input string parameter.
+func MockServerScanResp(t *testing.T, variant string) *ctlpb.StorageScanResp {
+	ssr := standardServerScanResponse(t)
+	switch variant {
+	case "withNamespace":
+		scmNamespaces := storage.ScmNamespaces{
+			storage.MockScmNamespace(),
+		}
+		if err := convert.Types(scmNamespaces, &ssr.Scm.Namespaces); err != nil {
+			t.Fatal(err)
+		}
+	case "noNVME":
+		ssr.Nvme.Ctrlrs = nil
+	case "noSCM":
+		ssr.Scm.Modules = nil
+	case "noStorage":
+		ssr.Nvme.Ctrlrs = nil
+		ssr.Scm.Modules = nil
+	case "scmFailed":
+		ssr.Scm.Modules = nil
+		ssr.Scm.State = &ctlpb.ResponseState{
+			Status: ctlpb.ResponseStatus_CTL_ERR_SCM,
+			Error:  "scm scan failed",
+		}
+	case "nvmeFailed":
+		ssr.Nvme.Ctrlrs = nil
+		ssr.Nvme.State = &ctlpb.ResponseState{
+			Status: ctlpb.ResponseStatus_CTL_ERR_NVME,
+			Error:  "nvme scan failed",
+		}
+	case "bothFailed":
+		ssr.Scm.Modules = nil
+		ssr.Scm.State = &ctlpb.ResponseState{
+			Status: ctlpb.ResponseStatus_CTL_ERR_SCM,
+			Error:  "scm scan failed",
+		}
+		ssr.Nvme.Ctrlrs = nil
+		ssr.Nvme.State = &ctlpb.ResponseState{
+			Status: ctlpb.ResponseStatus_CTL_ERR_NVME,
+			Error:  "nvme scan failed",
+		}
+	case "standard":
+	default:
+		t.Fatalf("MockServerScanResp(): variant %s unrecognised", variant)
+	}
+	return ssr
+}
+
+// MockHostResponses returns mock host responses.
+func MockHostResponses(t *testing.T, count int, fmtStr string, respMsg proto.Message) []*HostResponse {
+	hrs := make([]*HostResponse, count)
+	for i := 0; i < count; i++ {
+		hrs[i] = &HostResponse{
+			Addr:    fmt.Sprintf(fmtStr, i),
+			Message: respMsg,
+		}
+	}
+	return hrs
+}
+
+// MockFailureMap returns failure map from the given range of integers.
+func MockFailureMap(idxList ...int) map[int]struct{} {
+	fm := make(map[int]struct{})
+	for _, i := range idxList {
+		fm[i] = struct{}{}
+	}
+	return fm
+}
+
+// MockFormatConf configures the contents of a StorageFormatResp.
+type MockFormatConf struct {
+	Hosts        int
+	ScmPerHost   int
+	NvmePerHost  int
+	ScmFailures  map[int]struct{}
+	NvmeFailures map[int]struct{}
+}
+
+// MockFormatResp returns a populated StorageFormatResp based on input config.
+func MockFormatResp(t *testing.T, mfc MockFormatConf) *StorageFormatResp {
+	hem := make(HostErrorsMap)
+	hsm := make(HostStorageMap)
+
+	for i := 0; i < mfc.Hosts; i++ {
+		hs := &HostStorage{}
+		hostName := fmt.Sprintf("host%d", i+1)
+
+		for j := 0; j < mfc.ScmPerHost; j++ {
+			if _, failed := mfc.ScmFailures[j]; failed {
+				if err := hem.Add(hostName, errors.Errorf("/mnt/%d format failed", j+1)); err != nil {
+					t.Fatal(err)
+				}
+				continue
+			}
+			hs.ScmMountPoints = append(hs.ScmMountPoints, &storage.ScmMountPoint{
+				Info: ctlpb.ResponseStatus_CTL_SUCCESS.String(),
+				Path: fmt.Sprintf("/mnt/%d", j+1),
+			})
+		}
+
+		for j := 0; j < mfc.NvmePerHost; j++ {
+			if _, failed := mfc.NvmeFailures[j]; failed {
+				if err := hem.Add(hostName, errors.Errorf("NVMe device %d format failed", j+1)); err != nil {
+					t.Fatal(err)
+				}
+				continue
+			}
+
+			// If the SCM format/mount failed for this idx, then there shouldn't
+			// be an NVMe format result.
+			if _, failed := mfc.ScmFailures[j]; failed {
+				continue
+			}
+			hs.NvmeDevices = append(hs.NvmeDevices, &storage.NvmeController{
+				Info:    ctlpb.ResponseStatus_CTL_SUCCESS.String(),
+				PciAddr: fmt.Sprintf("%d", j+1),
+			})
+		}
+		if err := hsm.Add(hostName, hs); err != nil {
+			t.Fatal(err)
+		}
+	}
+
+	if len(hem) == 0 {
+		hem = nil
+	}
+	return &StorageFormatResp{
+		HostErrorsResp: HostErrorsResp{
+			HostErrors: hem,
+		},
+		HostStorage: hsm,
+	}
 }
