+++ src/control/lib/control/system_test.go
@@ -59,7 +59,7 @@ func TestControl_StartRanks(t *testing.T) {
 				},
 			},
 			expResp: &RanksResp{
-				HostErrorsResp: mockHostErrorsResp(t, &mockHostError{"host1", "remote failed"}),
+				HostErrorsResp: MockHostErrorsResp(t, &MockHostError{"host1", "remote failed"}),
 			},
 		},
 		"no results": {
@@ -116,7 +116,7 @@ func TestControl_StartRanks(t *testing.T) {
 					{Rank: 2, Action: "start", State: system.MemberStateReady},
 					{Rank: 3, Action: "start", Errored: true, Msg: "uh oh", State: system.MemberStateStopped},
 				},
-				HostErrorsResp: mockHostErrorsResp(t, &mockHostError{"host3", "connection refused"}),
+				HostErrorsResp: MockHostErrorsResp(t, &MockHostError{"host3", "connection refused"}),
 			},
 		},
 	} {
@@ -161,7 +161,7 @@ func TestControl_PrepShutdownRanks(t *testing.T) {
 				},
 			},
 			expResp: &RanksResp{
-				HostErrorsResp: mockHostErrorsResp(t, &mockHostError{"host1", "remote failed"}),
+				HostErrorsResp: MockHostErrorsResp(t, &MockHostError{"host1", "remote failed"}),
 			},
 		},
 		"no results": {
@@ -218,7 +218,7 @@ func TestControl_PrepShutdownRanks(t *testing.T) {
 					{Rank: 2, Action: "prep shutdown", State: system.MemberStateStopping},
 					{Rank: 3, Action: "prep shutdown", Errored: true, Msg: "uh oh", State: system.MemberStateStopped},
 				},
-				HostErrorsResp: mockHostErrorsResp(t, &mockHostError{"host3", "connection refused"}),
+				HostErrorsResp: MockHostErrorsResp(t, &MockHostError{"host3", "connection refused"}),
 			},
 		},
 	} {
@@ -263,7 +263,7 @@ func TestControl_StopRanks(t *testing.T) {
 				},
 			},
 			expResp: &RanksResp{
-				HostErrorsResp: mockHostErrorsResp(t, &mockHostError{"host1", "remote failed"}),
+				HostErrorsResp: MockHostErrorsResp(t, &MockHostError{"host1", "remote failed"}),
 			},
 		},
 		"no results": {
@@ -320,7 +320,7 @@ func TestControl_StopRanks(t *testing.T) {
 					{Rank: 2, Action: "stop", State: system.MemberStateStopped},
 					{Rank: 3, Action: "stop", Errored: true, Msg: "uh oh", State: system.MemberStateErrored},
 				},
-				HostErrorsResp: mockHostErrorsResp(t, &mockHostError{"host3", "connection refused"}),
+				HostErrorsResp: MockHostErrorsResp(t, &MockHostError{"host3", "connection refused"}),
 			},
 		},
 	} {
@@ -365,7 +365,7 @@ func TestControl_PingRanks(t *testing.T) {
 				},
 			},
 			expResp: &RanksResp{
-				HostErrorsResp: mockHostErrorsResp(t, &mockHostError{"host1", "remote failed"}),
+				HostErrorsResp: MockHostErrorsResp(t, &MockHostError{"host1", "remote failed"}),
 			},
 		},
 		"no results": {
@@ -422,7 +422,7 @@ func TestControl_PingRanks(t *testing.T) {
 					{Rank: 2, Action: "ping", State: system.MemberStateReady},
 					{Rank: 3, Action: "ping", Errored: true, Msg: "uh oh", State: system.MemberStateUnresponsive},
 				},
-				HostErrorsResp: mockHostErrorsResp(t, &mockHostError{"host3", "connection refused"}),
+				HostErrorsResp: MockHostErrorsResp(t, &MockHostError{"host3", "connection refused"}),
 			},
 		},
 	} {
@@ -980,7 +980,7 @@ func TestControl_SystemReformat(t *testing.T) {
 			),
 			expResp: &StorageFormatResp{
 				HostErrorsResp: HostErrorsResp{
-					HostErrors: mockHostErrorsMap(t, &mockHostError{
+					HostErrors: mockHostErrorsMap(t, &MockHostError{
 						"10.0.0.1:10001", "1 rank failed: didn't start",
 					}),
 				},
@@ -1042,9 +1042,9 @@ func TestControl_SystemReformat(t *testing.T) {
 			expResp: &StorageFormatResp{
 				HostErrorsResp: HostErrorsResp{
 					HostErrors: mockHostErrorsMap(t,
-						&mockHostError{"10.0.0.4:10001", "2 ranks failed: didn't start"},
-						&mockHostError{"10.0.0.[1,3]:10001", "1 rank failed: didn't start"},
-						&mockHostError{"10.0.0.3:10001", "1 rank failed: something bad"},
+						&MockHostError{"10.0.0.4:10001", "2 ranks failed: didn't start"},
+						&MockHostError{"10.0.0.[1,3]:10001", "1 rank failed: didn't start"},
+						&MockHostError{"10.0.0.3:10001", "1 rank failed: something bad"},
 					),
 				},
 			},
